requirements.txt
Django==3.2.8
gunicorn==20.1.0
azure-servicebus==7.3.0
